reformat: rename getClientCodeInPage to layoutBuild
